sqlproxyccl: move denylist files to acl package
Part of: https://cockroachlabs.atlassian.net/browse/CC-8136

Release note: None
