+++ pkg/BUILD.bazel
@@ -76,8 +76,8 @@ ALL_TESTS = [
     "//pkg/ccl/spanconfigccl/spanconfigsqlwatcherccl:spanconfigsqlwatcherccl_test",
     "//pkg/ccl/sqlitelogictestccl/tests/3node-tenant:3node-tenant_test",
     "//pkg/ccl/sqlitelogictestccl:sqlitelogictestccl_test",
+    "//pkg/ccl/sqlproxyccl/acl:acl_test",
     "//pkg/ccl/sqlproxyccl/balancer:balancer_test",
-    "//pkg/ccl/sqlproxyccl/denylist:denylist_test",
     "//pkg/ccl/sqlproxyccl/interceptor:interceptor_test",
     "//pkg/ccl/sqlproxyccl/tenant:tenant_test",
     "//pkg/ccl/sqlproxyccl/throttler:throttler_test",
@@ -814,10 +814,11 @@ GO_TARGETS = [
     "//pkg/ccl/sqlitelogictestccl/tests/3node-tenant:3node-tenant_test",
     "//pkg/ccl/sqlitelogictestccl:sqlitelogictestccl",
     "//pkg/ccl/sqlitelogictestccl:sqlitelogictestccl_test",
+    "//pkg/ccl/sqlproxyccl/acl:acl",
+    "//pkg/ccl/sqlproxyccl/acl:acl_test",
+    "//pkg/ccl/sqlproxyccl/acl:denylist",
     "//pkg/ccl/sqlproxyccl/balancer:balancer",
     "//pkg/ccl/sqlproxyccl/balancer:balancer_test",
-    "//pkg/ccl/sqlproxyccl/denylist:denylist",
-    "//pkg/ccl/sqlproxyccl/denylist:denylist_test",
     "//pkg/ccl/sqlproxyccl/interceptor:interceptor",
     "//pkg/ccl/sqlproxyccl/interceptor:interceptor_test",
     "//pkg/ccl/sqlproxyccl/tenant:tenant",
@@ -2419,8 +2420,8 @@ GET_X_DATA_TARGETS = [
     "//pkg/ccl/sqlitelogictestccl:get_x_data",
     "//pkg/ccl/sqlitelogictestccl/tests/3node-tenant:get_x_data",
     "//pkg/ccl/sqlproxyccl:get_x_data",
+    "//pkg/ccl/sqlproxyccl/acl:get_x_data",
     "//pkg/ccl/sqlproxyccl/balancer:get_x_data",
-    "//pkg/ccl/sqlproxyccl/denylist:get_x_data",
     "//pkg/ccl/sqlproxyccl/interceptor:get_x_data",
     "//pkg/ccl/sqlproxyccl/tenant:get_x_data",
     "//pkg/ccl/sqlproxyccl/tenantdirsvr:get_x_data",
