+++ pkg/ccl/sqlproxyccl/acl/denylist.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package denylist
+package acl
 
 import (
 	"time"
