+++ pkg/ccl/sqlproxyccl/acl/watcher_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package denylist
+package acl
 
 import (
 	"testing"
