+++ pkg/ccl/sqlproxyccl/proxy_handler_test.go
@@ -28,8 +28,8 @@ import (
 	"github.com/cockroachdb/cockroach-go/v2/crdb"
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvtenantccl"
+	"github.com/cockroachdb/cockroach/pkg/ccl/sqlproxyccl/acl"
 	"github.com/cockroachdb/cockroach/pkg/ccl/sqlproxyccl/balancer"
-	"github.com/cockroachdb/cockroach/pkg/ccl/sqlproxyccl/denylist"
 	"github.com/cockroachdb/cockroach/pkg/ccl/sqlproxyccl/tenant"
 	"github.com/cockroachdb/cockroach/pkg/ccl/sqlproxyccl/tenantdirsvr"
 	"github.com/cockroachdb/cockroach/pkg/ccl/sqlproxyccl/throttler"
@@ -667,7 +667,7 @@ func TestDenylistUpdate(t *testing.T) {
 	denyList, err := os.CreateTemp("", "*_denylist.yml")
 	require.NoError(t, err)
 	defer func() { _ = os.Remove(denyList.Name()) }()
-	dlf := denylist.File{Seq: 0}
+	dlf := acl.File{Seq: 0}
 	bytes, err := dlf.Serialize()
 	require.NoError(t, err)
 	_, err = denyList.Write(bytes)
@@ -741,9 +741,9 @@ func TestDenylistUpdate(t *testing.T) {
 
 	// Once connection has been established, attempt to update denylist.
 	dlf.Seq++
-	dlf.Denylist = []*denylist.DenyEntry{
+	dlf.Denylist = []*acl.DenyEntry{
 		{
-			Entity:     denylist.DenyEntity{Type: denylist.IPAddrType, Item: "127.0.0.1"},
+			Entity:     acl.DenyEntity{Type: acl.IPAddrType, Item: "127.0.0.1"},
 			Expiration: timeutil.Now().Add(time.Minute),
 			Reason:     "test-denied",
 		},
